added src and pom.xml

+++ src/main/java/in/bank/User.java
@@ -3,12 +3,13 @@ package in.bank;
 public class User {
 
 	public String name;
+	
 	public String email;
 	public String password;
 	public String address;
 	public long mobileNo;
 	public long accNo;
-	public long balance;
+	public double balance;
 	public boolean blockedAcc;
 	
 }
